testing/web-platform/tests/atomics/window-and-any-worker/wait/negative-timeout/negative-timeout-popup.html

<!doctype html>
<meta charset=utf-8>
<title></title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<script>

// Copyright (C) 2015 Mozilla Corporation.  All rights reserved.
// This code is governed by the BSD license found in the LICENSE file.

/*---
description: >
  Test that Atomics.wait times out with a negative timeout
---*/

window.addEventListener("load", function(e) {
  // Window->Worker->Worker.
  async_test(function(t) {
    let popup = window.open('../empty.html', 'empty');
    let test = this;
    popup.onmessage = function() {
      let worker = new Worker('wait-on-zero-negative.js');
      worker.onmessage = test.step_func_done(function(e) {
        assert_equals(e.data, "timed-out");
      });
      let ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
      worker.postMessage([ia.buffer]);
    }
    popup.postMessage([], '*');
  }, "atomics-dedicated-worker-wait-negative-timeout-from-popup")
})

</script>
